Eliminated lots of "size" warnings. Moved the free functions in cectorControl.cpp to VectorControlInternal.cpp.
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@12 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ include/intervention.h
@@ -26,7 +26,7 @@ public:
   
   /*! Common to all vaccine types. Number of vaccine doses that are given
    * either through EPI or as EPI Boosters. */
-  static int _numberOfEpiDoses;
+  static size_t _numberOfEpiDoses;
   
    /** Target age for EPI-like vaccination, in time steps
     * TODO: Should be an integer? */
